Update to recommended Xcode settings
The meaningful code changes here to deal with truncation warnings convert the uvc_range_t type to use long instead of int.

+++ UVCCameraControl.h
@@ -19,7 +19,7 @@
 #define UVC_GET_MAX	0x83
 
 typedef struct {
-	int min, max;
+	long min, max;
 } uvc_range_t;
 
 typedef struct {
